RDKAAMP-506 [AAMP] refactoring - change returned httpCode from long to int
Reason for change: code simplification
Test Procedure: regression only
Risks: low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_23_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_23_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_23_1

+++ AampEvent.cpp
@@ -856,7 +856,7 @@ bool BufferingChangedEvent::buffering() const
 /**
  * @brief DrmMetaDataEvent Constructor
  */
-DrmMetaDataEvent::DrmMetaDataEvent(AAMPTuneFailure failure, const std::string &accessStatus, int statusValue, long responseCode, bool secclientErr):
+DrmMetaDataEvent::DrmMetaDataEvent(AAMPTuneFailure failure, const std::string &accessStatus, int statusValue, int responseCode, bool secclientErr):
 	AAMPEventObject(AAMP_EVENT_DRM_METADATA), mFailure(failure), mAccessStatus(accessStatus),
 	mAccessStatusValue(statusValue), mResponseCode(responseCode), mSecclientError(secclientErr), mSecManagerReasonCode(-1), mSecManagerClass(-1),
 	mBusinessStatus(-1)
@@ -923,7 +923,7 @@ void DrmMetaDataEvent::setAccessStatusValue(int value)
  *
  * @return Response code
  */
-long DrmMetaDataEvent::getResponseCode() const
+int DrmMetaDataEvent::getResponseCode() const
 {
 	return mResponseCode;
 }
@@ -961,7 +961,7 @@ int32_t DrmMetaDataEvent::getBusinessStatus() const
 /**
  * @brief Set Response Code
  */
-void DrmMetaDataEvent::setResponseCode(long code)
+void DrmMetaDataEvent::setResponseCode(int code)
 {
 	mResponseCode = code;
 }
